user service get drafts

+++ api_service/models/forum.go
@@ -5,6 +5,6 @@ type Forum struct {
 	Slug    string `json:"slug,omitempty"`
 	Title   string `json:"title,omitempty"`
 	User    string `json:"user,omitempty"`
-	Posts   int64  `json:"posts,omitempty"`
-	Threads int64  `json:"threads,omitempty"`
+	Posts   int    `json:"posts,omitempty"`
+	Threads int    `json:"threads,omitempty"`
 }
