+++ api_service/models/post.go
@@ -4,14 +4,14 @@ import "time"
 
 //easyjson:json
 type Post struct {
-	ID       int64     `json:"id,omitempty"`
+	ID       int       `json:"id,omitempty"`
 	Message  string    `json:"message,omitempty"`
 	Created  time.Time `json:"created,omitempty"`
 	Author   string    `json:"author,omitempty"`
 	Forum    string    `json:"forum,omitempty"`
 	IsEdited bool      `json:"isEdited,omitempty"`
-	Parent   int64     `json:"parent,omitempty"`
-	Thread   int64     `json:"thread,omitempty"`
+	Parent   int       `json:"parent,omitempty"`
+	Thread   int       `json:"thread,omitempty"`
 }
 
 //easyjson:json
