+++ api_service/models/status.go
@@ -2,8 +2,8 @@ package models
 
 //easyjson:json
 type Status struct {
-	Forum  int64 `json:"forum,omitempty"`
-	Post   int64 `json:"post,omitempty"`
-	Thread int64 `json:"thread,omitempty"`
-	User   int64 `json:"user,omitempty"`
+	Forum  int `json:"forum,omitempty"`
+	Post   int `json:"post,omitempty"`
+	Thread int `json:"thread,omitempty"`
+	User   int `json:"user,omitempty"`
 }
