+++ api_service/models/vote.go
@@ -3,5 +3,5 @@ package models
 //easyjson:json
 type Vote struct {
 	Nickname string `json:"nickname,omitempty"`
-	Voice    int64  `json:"voice,omitempty"`
+	Voice    int    `json:"voice,omitempty"`
 }
